added

+++ src/driverapi/launcher/DynOpsClass.cpp
@@ -129,8 +129,8 @@ BPatch_function * DynOpsClass::FindFunctionInAddrList(BPatch_addressSpace * aspa
 			if (i->pathName().find(".so") != std::string::npos) {
 				std::vector<BPatch_module *> mods;
 				i->modules(mods);
-				if(_addressList.find(mods[0]->getBaseAddr() + p.libOffset) != _addressList.end())
-					return _addressList[mods[0]->getBaseAddr() + p.libOffset];
+				if(_addressList.find(((uint64_t)mods[0]->getBaseAddr()) + p.libOffset) != _addressList.end())
+					return _addressList[((uint64_t)mods[0]->getBaseAddr()) + p.libOffset];
 			}
 		}
 	}
